backend/project_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
pydantic==2.9.0
pyjwt==2.9.0
bcrypt==4.2.0
cryptography==43.0.0
httpx==0.27.0
google-cloud-iam==2.15.2
google-cloud-storage==2.18.2
google-auth==2.35.0
pytest==8.3.0
pytest-asyncio==0.24.0
testcontainers[postgres]==4.8.0
pyyaml==6.0.2
